# API_Neo4j/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API E-commerce Neo4j",
    "version": "1.0.0",
    "description": "API de ejemplo de e-commerce sobre Neo4j (usuarios, productos, pedidos y endpoints inseguros para laboratorio)."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Servidor local" }
  ],
  "paths": {
    "/usuarios": {
      "get": {
        "summary": "Listar usuarios",
        "tags": ["Usuarios"],
        "responses": {
          "200": { "description": "Lista de usuarios" }
        }
      },
      "post": {
        "summary": "Crear usuario",
        "tags": ["Usuarios"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "nombre": { "type": "string" },
                  "email": { "type": "string" },
                  "rol": { "type": "string" }
                },
                "required": ["id", "nombre", "email", "rol"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuario creado" }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "summary": "Obtener usuario por id",
        "tags": ["Usuarios"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Usuario" } }
      },
      "put": {
        "summary": "Actualizar usuario",
        "tags": ["Usuarios"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string" },
                  "email": { "type": "string" },
                  "rol": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Usuario actualizado" } }
      },
      "delete": {
        "summary": "Eliminar usuario",
        "tags": ["Usuarios"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "204": { "description": "Usuario eliminado" } }
      }
    },
    "/productos": {
      "get": {
        "summary": "Listar productos",
        "tags": ["Productos"],
        "responses": { "200": { "description": "Lista de productos" } }
      },
      "post": {
        "summary": "Crear producto",
        "tags": ["Productos"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "nombre": { "type": "string" },
                  "precio": { "type": "number" },
                  "stock": { "type": "integer" }
                },
                "required": ["id", "nombre", "precio", "stock"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Producto creado" } }
      }
    },
    "/productos/{id}": {
      "get": {
        "summary": "Obtener producto por id",
        "tags": ["Productos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Producto" } }
      },
      "put": {
        "summary": "Actualizar producto",
        "tags": ["Productos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string" },
                  "precio": { "type": "number" },
                  "stock": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Producto actualizado" } }
      },
      "delete": {
        "summary": "Eliminar producto",
        "tags": ["Productos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "204": { "description": "Producto eliminado" } }
      }
    },
    "/pedidos": {
      "get": {
        "summary": "Listar pedidos",
        "tags": ["Pedidos"],
        "responses": { "200": { "description": "Lista de pedidos con usuario y productos" } }
      },
      "post": {
        "summary": "Crear pedido",
        "tags": ["Pedidos"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "clienteId": { "type": "integer" },
                  "productos": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "cantidad": { "type": "integer" },
                        "precio_unitario": { "type": "number" }
                      }
                    }
                  }
                },
                "required": ["id", "clienteId", "productos"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Pedido creado" } }
      }
    },
    "/pedidos/{id}": {
      "get": {
        "summary": "Obtener pedido por id",
        "tags": ["Pedidos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Pedido con usuario y productos" } }
      },
      "put": {
        "summary": "Actualizar estado de pedido",
        "tags": ["Pedidos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "estado": { "type": "string" }
                },
                "required": ["estado"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Pedido actualizado" } }
      },
      "delete": {
        "summary": "Eliminar pedido",
        "tags": ["Pedidos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "204": { "description": "Pedido eliminado" } }
      }
    },
    "/pedidos/{id}/actualizar-estado": {
      "post": {
        "summary": "Actualizar estado de pedido (con rol)",
        "tags": ["Pedidos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "estado": { "type": "string" },
                  "rol": { "type": "string" }
                },
                "required": ["estado", "rol"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Pedido actualizado" },
          "403": { "description": "No autorizado" }
        }
      }
    },
    "/buscar-productos": {
      "get": {
        "summary": "Buscar productos por nombre (CONTAINS)",
        "tags": ["Búsquedas"],
        "parameters": [
          { "name": "nombre", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Productos coincidentes" } }
      }
    },
    "/buscar-usuarios-inseguro": {
      "get": {
        "summary": "Búsqueda insegura de usuarios por nombre",
        "tags": ["Inseguros"],
        "parameters": [
          { "name": "nombre", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Usuarios coincidentes (consulta sin parámetros)" } }
      }
    },
    "/login-inseguro": {
      "get": {
        "summary": "Login inseguro de usuario fijo",
        "tags": ["Inseguros"],
        "parameters": [
          { "name": "password", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Usuario si la contraseña coincide" } }
      }
    }
  }
}